tests/wpt/tests/mediacapture-streams/MediaStream-supported-by-permissions-policy.html

<!DOCTYPE html>
<title>Test that camera and microphone are advertised in the feature list</title>
<link rel="help" href="https://w3c.github.io/webappsec-permissions-policy/#dom-permissionspolicy-features">
<link rel="help" href="https://w3c.github.io/mediacapture-main/#permissions-policy-integration">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
test(() => {
    assert_in_array('camera', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise camera.');

test(() => {
    assert_in_array('microphone', document.permissionsPolicy.features());
}, 'document.permissionsPolicy.features should advertise microphone.');
</script>
